+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Li 2024 ceftazidime ([doi:10.3389/fphar.2024.1470350](https://doi.org/10.3389/fphar.2024.1470350)) -- critically ill children in a paediatric intensive care unit.
+- Add Axelsen 2024 selexipag adult and pediatric population PK models ([doi:10.1002/psp4.13231](https://doi.org/10.1002/psp4.13231)) -- adults with pulmonary arterial hypertension and pediatric patients aged 2 to <18 years.
 
 - Add Sager 2023 sotrovimab population PK and COVID-19 progression exposure-response models ([doi:10.1002/psp4.12958](https://doi.org/10.1002/psp4.12958)) -- non-hospitalized adults and adolescents with mild-to-moderate COVID-19 at high risk of progression, plus healthy volunteers.